CSC4730/6730, CMIS 4200 Data Visualization
CSC4730/6730, CMIS 4200
Data Visualization

Summer 2018

Project 2

Due date: 11:59 pm, July 6, 2018

Goal: 
1.	Learn how to display data on a Google Map using a Google Fusion Table layer.
2.	Learn how to use Google Place Search along with Google Maps. 
3.	Learn how to use the Google Maps Direction service. 
4.	Learn how to display Google Street View along with Google Maps. 

Requirements:

1.	Create a web page with four sections: A, B, C, and D. 
2.	(20 points) Section A should contain two input boxes: Address and Place. 
a.	In the “Address” box, a user can enter a street address, city, or a zip code. 
b.	In the “Place” search box, a user can enter the type or the name of a business or place (e.g. restaurant, hotel, hospital, or Home Depot). Use Google Places service to conduct the search around the address entered into the “Address” box. and display the results in section B (Google Maps view). 
c.	There must be a button called “submit”. 
3.	(40 points) Section B should be a Google Maps view.
a.	At the beginning, before a user enters anything and press the “submit” button in section A, section B (Google Maps view) should visualize a data set stored in a Google Fusion Table. 
i.	You may use the following Fusion Tables. 
1.	Geocoded Wikipedia articles
2.	Northampton housing price
3.	Atlanta Community Resources
4.	Or create your own Fusion Table with your own data. Make sure your data includes latitude/longitude or street address. The Fusion Table you use must be “Public” or “Anyone with the link can view”.

ii.	Each row of the Fusion Table must contain one or two location column(s). 
iii.	Each row of the Fusion Table must be displayed as a marker on the map. 
iv.	When a user clicks on a maker, an information window shall be displayed, showing at least two columns of data from that row. 



b.	Once a user enters information in the search boxes in section A and press the submit button, the Google Maps view should center on the address entered in the “Address” box in section A. 
i.	If no address is entered, then use the previous location. 

c.	The Google Maps view should display the places (or businesses) returned by Google Places search based on the keywords entered in the “Place” search box. 

d.	When a user selects a marker on the map, the driving directions from the address in the “Address” box (in section A) to the location of the clicked marker (e.g. a restaurant) should be displayed in section C. In the same time, the Google Street View for the selected location should be displayed in section D. 

i.	This applies to bother the Google Fusion Table markers and the markers from Google Places search. 

ii.	If no address is entered, display a message asking user to enter an address. 

4.	(20 points) Section C should display the driving direction from the user entered address (in section A) to the user selected icon on the map (in section B).

5.	(20 points) Section D should display the Google Street View of the selected marker.  

6.	Implement your program with Google Maps JavaScript API. 

7.	This is an individual project. Every student must submit his/her own work.

Grading guidelines:
•	Your program must work “out of the box”. My TA or I will just copy and paste your code to jsbin.com (or another JavaScript playground site) to test your program. You can easily test it yourself. 
•	The program must meet all the requirements above. Points will be deducted for missing requirements. 
•	Points will be deducted if the program has no or very few comments. 
•	Late submissions may be accepted with the following conditions:
o	If it’s 1 or 2 days late from the posted deadline, a 30% penalty will be imposed. In other words, the highest grade you will possibly get is 70 out of 100. 
o	If it’s 3 to 6 days late from the posted deadline, a 50% penalty will be imposed. In other words, the highest grade you will possibly get is 50 out of 100.
o	Projects submitted more than 6 days late will not be accepted. 

Deliverables: 
•	Put your HTML/JavaScript file in a ZIP file and submit it to iCollege, under folder Project 2. 
•	Do not submit the HTML file directly. For some reason, when you retrieve a HTML file from iCollege, the file may be corrupted and unusable.  
•	Do not submit by email.